Moved all sources inside kernel
