+++ CMakeLists.txt
@@ -4,12 +4,6 @@ project(bolgenos-ng
 	VERSION 0.0.4
 	LANGUAGES NONE)
 
-set(kernel_headers_dir "include/")
-include_directories(${kernel_headers_dir})
-
-set_source_files_properties(${kernel_std_headers} PROPERTIES LANGUAGE CXX)
-set(kernel_headers ${kernel_std_headers} ${kernel_headers})
-
 if (${CMAKE_CROSSCOMPILING})
 	message(STATUS "CMake is being run in crosscompiling mode. "
 		"Kernel targets will be built.")
