--- include/bolgenos-ng/log.hpp
@@ -1,1 +1,0 @@
-../../kernel/log/include/log.hpp
