--- include/bolgenos-ng/loggable.hpp
@@ -1,1 +1,0 @@
-../../kernel/log/include/loggable.hpp
