+++ kernel/devices/include/bolgenos-ng/device.hpp
