+++ kernel/devices/include/bolgenos-ng/vga_console.hpp
