--- kernel/devices/libdevices.cpp
@@ -1,2 +1,0 @@
-
-// an empty source to make CMake happy
