+++ kernel/devices/src/frequency_divider.hpp
