+++ kernel/devices/src/keyboard.cpp
@@ -6,11 +6,11 @@ namespace {
 
 using namespace ps2::keyboard;
 
-int __is_letter(ps2::keyboard::kb_key key) {
+int is_letter(ps2::keyboard::kb_key key) {
 	return (kb_key_a <= key && key <= kb_key_z);
 }
 
-int __is_digit(ps2::keyboard::kb_key key) {
+int is_digit(ps2::keyboard::kb_key key) {
 	return (kb_key_0 <= key && key <= kb_key_9);
 }
 
@@ -26,11 +26,11 @@ ps2::keyboard::kb_key& ps2::keyboard::operator++(kb_key &key) {
 
 
 char ps2::keyboard::to_printable_key(kb_key key) {
-	if (__is_letter(key)) {
+	if (is_letter(key)) {
 		return 'a' + (key - kb_key_a);
 	}
 
-	if (__is_digit(key)) {
+	if (is_digit(key)) {
 		return '0' + (key - kb_key_0);
 	}
