+++ kernel/devices/src/pit.cpp
@@ -7,7 +7,7 @@
 #include <bolgenos-ng/interrupt_controller.hpp>
 #include <bolgenos-ng/irq.hpp>
 #include <bolgenos-ng/log.hpp>
-#include <bolgenos-ng/mem_utils.hpp>
+#include <mem_utils.hpp>
 #include <bolgenos-ng/time.hpp>
 
 #include "frequency_divider.hpp"
