+++ kernel/libkernelcxx/include/bitarray.hpp
