+++ kernel/libkernelcxx/include/bolgenos-ng/compiler.h
