+++ kernel/libkernelcxx/include/bolgenos-ng/error.h
