+++ kernel/libkernelcxx/include/bolgenos-ng/printk.h
