+++ kernel/libkernelcxx/include/bolgenos-ng/time.hpp
