+++ kernel/libkernelcxx/include/mem_utils.hpp
