+++ kernel/libkernelcxx/src/compiler.cpp
