+++ kernel/log/include/bolgenos-ng/log.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "../log.hpp"
