+++ kernel/log/include/bolgenos-ng/loggable.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "../loggable.hpp"
