+++ kernel/x86/include/x86/gate.hpp
@@ -4,7 +4,7 @@
 #include "segment_flags.hpp"
 #include "gdt.hpp"
 
-#include <bolgenos-ng/mem_utils.hpp>
+#include <mem_utils.hpp>
 
 namespace x86 {
